update hangman to stage 7/8 no completed yet
If the user enters the same letter twice then the program should output You already typed this letter .
Also, you should check if the user prints an English lowercase letter or not. If not, the program should print It is not an ASCII lowercase letter .
Also, you should check if the user printed exactly one letter. If not, the program should print You should print a single letter . Remember that zero is also not one!
Note that all these three errors should not reduce attempts count!